move hooks to hooks and x402 lib to subfolder
