+++ src/lib/x402/schema.ts
